Combine categorical variables into one

Combine specified categorical variables by concatenating their values into one character, and returns the result along with tidyverse code used to generate it.

combineCatVars(
  .data,
  vars,
  sep = ".",
  name = paste(vars, collapse = sep),
  keep_empty = FALSE
)

Arguments

.data
a dataframe with the columns to be combined

vars
a character vector of the categorical variables to be combined

sep
the separator to combine the values of the variables in var by. "." by default

name
a name for the new variable

keep_empty
logical, if FALSE empty level combinations are removed from the factor

Value

original dataframe containing a new column of the renamed categorical variable with tidyverse code attached

Details

When either variable is NA, the result is NA.

Author

Owen Jin

Examples

combined <- combineCatVars(warpbreaks, vars = c("wool", "tension"), sep = "_")
cat(code(combined))
#> warpbreaks %>% dplyr::mutate(wool_tension = forcats::fct_cross(wool,      tension, sep = "_", keep_empty = FALSE))
head(combined)
#>   breaks wool tension wool_tension
#> 1     26    A       L          A_L
#> 2     30    A       L          A_L
#> 3     54    A       L          A_L
#> 4     25    A       L          A_L
#> 5     70    A       L          A_L
#> 6     52    A       L          A_L