Create new variables

Create a new variable by using a valid R expression and returns the result along with tidyverse code used to generate it.

createNewVar(.data, new_var = "new.variable", R_exp)

Arguments

.data

a dataframe to which to add a new variable to

new_var

a character of the new variable name. "new.variable" by default

R_exp

a character of a valid R expression which can generate a vector of values

Value

original dataframe containing the new column created from R_exp

with tidyverse code attached

See also

Author

Owen Jin

Examples

created <- createNewVar(iris, new_var = "Sepal.Length_less_Sepal.Width",
 "Sepal.Length - Sepal.Width")
cat(code(created))
#> iris %>% dplyr::mutate(Sepal.Length_less_Sepal.Width = Sepal.Length -      Sepal.Width)
head(created)
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa
#> 6          5.4         3.9          1.7         0.4  setosa
#>   Sepal.Length_less_Sepal.Width
#> 1                           1.6
#> 2                           1.9
#> 3                           1.5
#> 4                           1.5
#> 5                           1.4
#> 6                           1.5