Filter data by levels of a numeric variables

Filter a dataframe by some boolean condition of one numeric variable and returns the result along with tidyverse code used to generate it.

filterNumeric(.data, var, op, num)

Arguments

.data
a dataframe or survey design object to filter

var
character of the column in .data to filter by

op
a logical operator of "<=", "<", ">=", ">", "==" or "!=" for the boolean condition

num
a number for which the op applies to

Value

filtered dataframe with tidyverse code attached

See also

Author

Owen Jin, Tom Elliott

Examples

filtered <- filterNumeric(iris, var = "Sepal.Length", op = "<=", num = 5)
cat(code(filtered))
#> iris %>% dplyr::filter(Sepal.Length <= 5)
head(filtered)
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          4.9         3.0          1.4         0.2  setosa
#> 2          4.7         3.2          1.3         0.2  setosa
#> 3          4.6         3.1          1.5         0.2  setosa
#> 4          5.0         3.6          1.4         0.2  setosa
#> 5          4.6         3.4          1.4         0.3  setosa
#> 6          5.0         3.4          1.5         0.2  setosa

require(survey)
#> Loading required package: survey
#> Loading required package: grid
#> Loading required package: Matrix
#> Loading required package: survival
#> 
#> Attaching package: ‘survey’
#> The following object is masked from ‘package:graphics’:
#> 
#>     dotchart
data(api)
svy <- svydesign(~dnum+snum, weights = ~pw, fpc = ~fpc1+fpc2, data = apiclus2)
(svy_filtered <- filterNumeric(svy, var = "api00", op = "<", num = 700))
#> 2 - level Cluster Sampling design
#> With (24, 63) clusters.
#> Called via srvyr
#> Data variables: cds (chr), stype (fct), name (chr), sname (chr), snum (dbl),
#>   dname (chr), dnum (int), cname (chr), cnum (int), flag (int), pcttest (int),
#>   api00 (int), api99 (int), target (int), growth (int), sch.wide (fct),
#>   comp.imp (fct), both (fct), awards (fct), meals (int), ell (int), yr.rnd
#>   (fct), mobility (int), acs.k3 (int), acs.46 (int), acs.core (int), pct.resp
#>   (int), not.hsg (int), hsg (int), some.col (int), col.grad (int), grad.sch
#>   (int), avg.ed (dbl), full (int), emer (int), enroll (int), api.stu (int), pw
#>   (dbl), fpc1 (dbl), fpc2 (int)
cat(code(svy_filtered))
#> svy %>% srvyr::as_survey() %>% srvyr::filter(api00 < 700)