Form Class Intervals

Create categorical intervals from a numeric variable.

form_class_intervals(
  .data,
  variable,
  method = c("equal", "width", "count", "manual"),
  n_intervals = 4L,
  interval_width,
  format = "(a,b]",
  range = NULL,
  format.lowest = ifelse(isinteger, "< a", "<= a"),
  format.highest = "> b",
  break_points = NULL,
  name = sprintf("%s.f", variable)
)

Arguments

.data

the data set

variable

name of the variable to convert

method

one of 'equal' for equal-width intervals, 'width' for intervals of a specific width, 'count' for equal-count intervals, and 'manual' to specify break points manually

n_intervals

for methods 'equal' and 'count', this is the number of intervals to create

interval_width

for method 'width', this is the width of intervals

format

the format for intervals; use 'a' and 'b' to represent the min/max of each interval, respectively.

range

the range of the data; use this to adjust the labels (e.g., for continuous data, set this to floor/ceiling of the min/max of the data to get prettier intervals). If range does not cover the range of the data, values outside will be placed into 'less than a' and 'greater than b' categories

format.lowest

values lower than the min of range will have this label format

format.highest

values higher than the max of range will have this label format

break_points

for method 'manual', specify breakpoints here (as a numeric vector)

name

the name of the new variable in the resulting data set

Value

a dataframe with an additional column with categorical class intervals

Author

Tom Elliott

Examples

form_class_intervals(iris, 'Sepal.Length', 'equal', 5L)
#>     Sepal.Length Sepal.Length.f Sepal.Width Petal.Length Petal.Width    Species
#> 1            5.1      (4.8,5.6]         3.5          1.4         0.2     setosa
#> 2            4.9      (4.8,5.6]         3.0          1.4         0.2     setosa
#> 3            4.7        [4,4.8]         3.2          1.3         0.2     setosa
#> 4            4.6        [4,4.8]         3.1          1.5         0.2     setosa
#> 5            5.0      (4.8,5.6]         3.6          1.4         0.2     setosa
#> 6            5.4      (4.8,5.6]         3.9          1.7         0.4     setosa
#> 7            4.6        [4,4.8]         3.4          1.4         0.3     setosa
#> 8            5.0      (4.8,5.6]         3.4          1.5         0.2     setosa
#> 9            4.4        [4,4.8]         2.9          1.4         0.2     setosa
#> 10           4.9      (4.8,5.6]         3.1          1.5         0.1     setosa
#> 11           5.4      (4.8,5.6]         3.7          1.5         0.2     setosa
#> 12           4.8        [4,4.8]         3.4          1.6         0.2     setosa
#> 13           4.8        [4,4.8]         3.0          1.4         0.1     setosa
#> 14           4.3        [4,4.8]         3.0          1.1         0.1     setosa
#> 15           5.8      (5.6,6.4]         4.0          1.2         0.2     setosa
#> 16           5.7      (5.6,6.4]         4.4          1.5         0.4     setosa
#> 17           5.4      (4.8,5.6]         3.9          1.3         0.4     setosa
#> 18           5.1      (4.8,5.6]         3.5          1.4         0.3     setosa
#> 19           5.7      (5.6,6.4]         3.8          1.7         0.3     setosa
#> 20           5.1      (4.8,5.6]         3.8          1.5         0.3     setosa
#> 21           5.4      (4.8,5.6]         3.4          1.7         0.2     setosa
#> 22           5.1      (4.8,5.6]         3.7          1.5         0.4     setosa
#> 23           4.6        [4,4.8]         3.6          1.0         0.2     setosa
#> 24           5.1      (4.8,5.6]         3.3          1.7         0.5     setosa
#> 25           4.8        [4,4.8]         3.4          1.9         0.2     setosa
#> 26           5.0      (4.8,5.6]         3.0          1.6         0.2     setosa
#> 27           5.0      (4.8,5.6]         3.4          1.6         0.4     setosa
#> 28           5.2      (4.8,5.6]         3.5          1.5         0.2     setosa
#> 29           5.2      (4.8,5.6]         3.4          1.4         0.2     setosa
#> 30           4.7        [4,4.8]         3.2          1.6         0.2     setosa
#> 31           4.8        [4,4.8]         3.1          1.6         0.2     setosa
#> 32           5.4      (4.8,5.6]         3.4          1.5         0.4     setosa
#> 33           5.2      (4.8,5.6]         4.1          1.5         0.1     setosa
#> 34           5.5      (4.8,5.6]         4.2          1.4         0.2     setosa
#> 35           4.9      (4.8,5.6]         3.1          1.5         0.2     setosa
#> 36           5.0      (4.8,5.6]         3.2          1.2         0.2     setosa
#> 37           5.5      (4.8,5.6]         3.5          1.3         0.2     setosa
#> 38           4.9      (4.8,5.6]         3.6          1.4         0.1     setosa
#> 39           4.4        [4,4.8]         3.0          1.3         0.2     setosa
#> 40           5.1      (4.8,5.6]         3.4          1.5         0.2     setosa
#> 41           5.0      (4.8,5.6]         3.5          1.3         0.3     setosa
#> 42           4.5        [4,4.8]         2.3          1.3         0.3     setosa
#> 43           4.4        [4,4.8]         3.2          1.3         0.2     setosa
#> 44           5.0      (4.8,5.6]         3.5          1.6         0.6     setosa
#> 45           5.1      (4.8,5.6]         3.8          1.9         0.4     setosa
#> 46           4.8        [4,4.8]         3.0          1.4         0.3     setosa
#> 47           5.1      (4.8,5.6]         3.8          1.6         0.2     setosa
#> 48           4.6        [4,4.8]         3.2          1.4         0.2     setosa
#> 49           5.3      (4.8,5.6]         3.7          1.5         0.2     setosa
#> 50           5.0      (4.8,5.6]         3.3          1.4         0.2     setosa
#> 51           7.0      (6.4,7.2]         3.2          4.7         1.4 versicolor
#> 52           6.4      (5.6,6.4]         3.2          4.5         1.5 versicolor
#> 53           6.9      (6.4,7.2]         3.1          4.9         1.5 versicolor
#> 54           5.5      (4.8,5.6]         2.3          4.0         1.3 versicolor
#> 55           6.5      (6.4,7.2]         2.8          4.6         1.5 versicolor
#> 56           5.7      (5.6,6.4]         2.8          4.5         1.3 versicolor
#> 57           6.3      (5.6,6.4]         3.3          4.7         1.6 versicolor
#> 58           4.9      (4.8,5.6]         2.4          3.3         1.0 versicolor
#> 59           6.6      (6.4,7.2]         2.9          4.6         1.3 versicolor
#> 60           5.2      (4.8,5.6]         2.7          3.9         1.4 versicolor
#> 61           5.0      (4.8,5.6]         2.0          3.5         1.0 versicolor
#> 62           5.9      (5.6,6.4]         3.0          4.2         1.5 versicolor
#> 63           6.0      (5.6,6.4]         2.2          4.0         1.0 versicolor
#> 64           6.1      (5.6,6.4]         2.9          4.7         1.4 versicolor
#> 65           5.6      (4.8,5.6]         2.9          3.6         1.3 versicolor
#> 66           6.7      (6.4,7.2]         3.1          4.4         1.4 versicolor
#> 67           5.6      (4.8,5.6]         3.0          4.5         1.5 versicolor
#> 68           5.8      (5.6,6.4]         2.7          4.1         1.0 versicolor
#> 69           6.2      (5.6,6.4]         2.2          4.5         1.5 versicolor
#> 70           5.6      (4.8,5.6]         2.5          3.9         1.1 versicolor
#> 71           5.9      (5.6,6.4]         3.2          4.8         1.8 versicolor
#> 72           6.1      (5.6,6.4]         2.8          4.0         1.3 versicolor
#> 73           6.3      (5.6,6.4]         2.5          4.9         1.5 versicolor
#> 74           6.1      (5.6,6.4]         2.8          4.7         1.2 versicolor
#> 75           6.4      (5.6,6.4]         2.9          4.3         1.3 versicolor
#> 76           6.6      (6.4,7.2]         3.0          4.4         1.4 versicolor
#> 77           6.8      (6.4,7.2]         2.8          4.8         1.4 versicolor
#> 78           6.7      (6.4,7.2]         3.0          5.0         1.7 versicolor
#> 79           6.0      (5.6,6.4]         2.9          4.5         1.5 versicolor
#> 80           5.7      (5.6,6.4]         2.6          3.5         1.0 versicolor
#> 81           5.5      (4.8,5.6]         2.4          3.8         1.1 versicolor
#> 82           5.5      (4.8,5.6]         2.4          3.7         1.0 versicolor
#> 83           5.8      (5.6,6.4]         2.7          3.9         1.2 versicolor
#> 84           6.0      (5.6,6.4]         2.7          5.1         1.6 versicolor
#> 85           5.4      (4.8,5.6]         3.0          4.5         1.5 versicolor
#> 86           6.0      (5.6,6.4]         3.4          4.5         1.6 versicolor
#> 87           6.7      (6.4,7.2]         3.1          4.7         1.5 versicolor
#> 88           6.3      (5.6,6.4]         2.3          4.4         1.3 versicolor
#> 89           5.6      (4.8,5.6]         3.0          4.1         1.3 versicolor
#> 90           5.5      (4.8,5.6]         2.5          4.0         1.3 versicolor
#> 91           5.5      (4.8,5.6]         2.6          4.4         1.2 versicolor
#> 92           6.1      (5.6,6.4]         3.0          4.6         1.4 versicolor
#> 93           5.8      (5.6,6.4]         2.6          4.0         1.2 versicolor
#> 94           5.0      (4.8,5.6]         2.3          3.3         1.0 versicolor
#> 95           5.6      (4.8,5.6]         2.7          4.2         1.3 versicolor
#> 96           5.7      (5.6,6.4]         3.0          4.2         1.2 versicolor
#> 97           5.7      (5.6,6.4]         2.9          4.2         1.3 versicolor
#> 98           6.2      (5.6,6.4]         2.9          4.3         1.3 versicolor
#> 99           5.1      (4.8,5.6]         2.5          3.0         1.1 versicolor
#> 100          5.7      (5.6,6.4]         2.8          4.1         1.3 versicolor
#> 101          6.3      (5.6,6.4]         3.3          6.0         2.5  virginica
#> 102          5.8      (5.6,6.4]         2.7          5.1         1.9  virginica
#> 103          7.1      (6.4,7.2]         3.0          5.9         2.1  virginica
#> 104          6.3      (5.6,6.4]         2.9          5.6         1.8  virginica
#> 105          6.5      (6.4,7.2]         3.0          5.8         2.2  virginica
#> 106          7.6        (7.2,8]         3.0          6.6         2.1  virginica
#> 107          4.9      (4.8,5.6]         2.5          4.5         1.7  virginica
#> 108          7.3        (7.2,8]         2.9          6.3         1.8  virginica
#> 109          6.7      (6.4,7.2]         2.5          5.8         1.8  virginica
#> 110          7.2      (6.4,7.2]         3.6          6.1         2.5  virginica
#> 111          6.5      (6.4,7.2]         3.2          5.1         2.0  virginica
#> 112          6.4      (5.6,6.4]         2.7          5.3         1.9  virginica
#> 113          6.8      (6.4,7.2]         3.0          5.5         2.1  virginica
#> 114          5.7      (5.6,6.4]         2.5          5.0         2.0  virginica
#> 115          5.8      (5.6,6.4]         2.8          5.1         2.4  virginica
#> 116          6.4      (5.6,6.4]         3.2          5.3         2.3  virginica
#> 117          6.5      (6.4,7.2]         3.0          5.5         1.8  virginica
#> 118          7.7        (7.2,8]         3.8          6.7         2.2  virginica
#> 119          7.7        (7.2,8]         2.6          6.9         2.3  virginica
#> 120          6.0      (5.6,6.4]         2.2          5.0         1.5  virginica
#> 121          6.9      (6.4,7.2]         3.2          5.7         2.3  virginica
#> 122          5.6      (4.8,5.6]         2.8          4.9         2.0  virginica
#> 123          7.7        (7.2,8]         2.8          6.7         2.0  virginica
#> 124          6.3      (5.6,6.4]         2.7          4.9         1.8  virginica
#> 125          6.7      (6.4,7.2]         3.3          5.7         2.1  virginica
#> 126          7.2      (6.4,7.2]         3.2          6.0         1.8  virginica
#> 127          6.2      (5.6,6.4]         2.8          4.8         1.8  virginica
#> 128          6.1      (5.6,6.4]         3.0          4.9         1.8  virginica
#> 129          6.4      (5.6,6.4]         2.8          5.6         2.1  virginica
#> 130          7.2      (6.4,7.2]         3.0          5.8         1.6  virginica
#> 131          7.4        (7.2,8]         2.8          6.1         1.9  virginica
#> 132          7.9        (7.2,8]         3.8          6.4         2.0  virginica
#> 133          6.4      (5.6,6.4]         2.8          5.6         2.2  virginica
#> 134          6.3      (5.6,6.4]         2.8          5.1         1.5  virginica
#> 135          6.1      (5.6,6.4]         2.6          5.6         1.4  virginica
#> 136          7.7        (7.2,8]         3.0          6.1         2.3  virginica
#> 137          6.3      (5.6,6.4]         3.4          5.6         2.4  virginica
#> 138          6.4      (5.6,6.4]         3.1          5.5         1.8  virginica
#> 139          6.0      (5.6,6.4]         3.0          4.8         1.8  virginica
#> 140          6.9      (6.4,7.2]         3.1          5.4         2.1  virginica
#> 141          6.7      (6.4,7.2]         3.1          5.6         2.4  virginica
#> 142          6.9      (6.4,7.2]         3.1          5.1         2.3  virginica
#> 143          5.8      (5.6,6.4]         2.7          5.1         1.9  virginica
#> 144          6.8      (6.4,7.2]         3.2          5.9         2.3  virginica
#> 145          6.7      (6.4,7.2]         3.3          5.7         2.5  virginica
#> 146          6.7      (6.4,7.2]         3.0          5.2         2.3  virginica
#> 147          6.3      (5.6,6.4]         2.5          5.0         1.9  virginica
#> 148          6.5      (6.4,7.2]         3.0          5.2         2.0  virginica
#> 149          6.2      (5.6,6.4]         3.4          5.4         2.3  virginica
#> 150          5.9      (5.6,6.4]         3.0          5.1         1.8  virginica