Import linked data into an inzdf object

Import linked data into an inzdf object

Usage

load_linked(
  x,
  schema,
  con,
  name = ifelse(missing(con), deparse(substitute(x)), deparse(substitute(con))),
  keep_con = FALSE,
  progress = FALSE,
  ...
)

Arguments

x

a linked specification file or vector of data set paths

schema

a list describing the schema/relationships between the files

con

a database connection to load the linked data into

name

the name of the data set collection

keep_con

if TRUE data will remain in DB (use for very large data)

progress

either TRUE or FALSE to enable/disable the default progress bar, or a list of three functions to x <- create(from, to), set(x, i), and destroy(x) a progress bar.

...

additional arguments passed to data reading function smart_read()

Value

an inzdf object