Read CSV with iNZight metadata

This function will read a CSV file with iNZight metadata in the header. This allows plain text CSV files to be supplied with additional comments that describe the structure of the data to make import and data handling easier.

Usage

read_meta(file, preview = FALSE, column_types, ...)

Arguments

file

the plain text file with metadata

preview

logical, if TRUE only the first 10 rows are returned

column_types

optional column types

...

more arguments

Value

a data frame

Details

The main example is to define factor levels for an integer variable in large data sets.

Author

Tom Elliott