Reorder a categorical

Reorder the factors of a categorical variable either manually or frequency

reorderLevels(
  .data,
  var,
  new_levels = NULL,
  freq = FALSE,
  name = sprintf("%s.reord", var)
)

Arguments

.data
a dataframe to reorder

var
a categorical variable to reorder

new_levels
a character vector of the new factor order. Only specify if freq = FALSE

freq
logical, If freq = FALSE (default), will manually reorder using new_levels. If freq = TRUE, will reorder based of descending frequency of the factor levels

name
name for the new variable

Value

original dataframe containing a new column of the reordered categorical variable with tidyverse code attached

See also

Author

Owen Jin

Examples

reordered <- reorderLevels(iris, var = "Species",
    new_levels = c("versicolor", "virginica", "setosa"))
cat(code(reordered))
#> iris %>% tibble::add_column(Species.reord = factor(iris$Species,      levels = c("versicolor", "virginica", "setosa")), .after = "Species")
head(reordered)
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species Species.reord
#> 1          5.1         3.5          1.4         0.2  setosa        setosa
#> 2          4.9         3.0          1.4         0.2  setosa        setosa
#> 3          4.7         3.2          1.3         0.2  setosa        setosa
#> 4          4.6         3.1          1.5         0.2  setosa        setosa
#> 5          5.0         3.6          1.4         0.2  setosa        setosa
#> 6          5.4         3.9          1.7         0.4  setosa        setosa