iNZight Tidy Code

Tidy code with correct indents and limit the code to the specific width

Usage

tidy_all_code(x, width = 80, indent = 4, outfile, incl_library = TRUE)

Arguments

x

character string or file name of the file containing messy code

width

the width of a line

indent

how many spaces for one indent

outfile

the file name of the file containing formatted code

incl_library

logical, if true, the output code will contain library name

Value

formatted code, optionally written to `outfile`

Author

Tom Elliott, Lushi Cai