Transform data of numeric variables

Transform the values of numeric variables by applying a mathematical function

Usage

transform_vars(data, vars, fn, names = NULL)

Arguments

data

a dataframe with the variables to transform

vars

a character of the numeric variables in data to transform

fn

the name (a string) of a valid R function

names

the names of the new variables

Value

the original dataframe containing the new columns of the transformed variable with tidyverse code attached

See also

Author

Zhaoming Su

Examples

transformed <- transform_vars(iris,
    var = "Petal.Length",
    fn = "log"
)
cat(code(transformed))
#> iris |> dplyr::mutate(log.Petal.Length = log(Petal.Length), .after = Petal.Length)
head(transformed)
#>   Sepal.Length Sepal.Width Petal.Length log.Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4        0.3364722         0.2  setosa
#> 2          4.9         3.0          1.4        0.3364722         0.2  setosa
#> 3          4.7         3.2          1.3        0.2623643         0.2  setosa
#> 4          4.6         3.1          1.5        0.4054651         0.2  setosa
#> 5          5.0         3.6          1.4        0.3364722         0.2  setosa
#> 6          5.4         3.9          1.7        0.5306283         0.4  setosa