Collapse data by values of a categorical variable

Collapse several values in a categorical variable into one level

collapseLevels(
  .data,
  var,
  levels,
  collapse = paste(levels, collapse = "_"),
  name = sprintf("%s.coll", var)
)

Arguments

.data

a dataframe to collapse

var

a character of the name of the categorical variable to collapse

levels

a character vector of the levels to be collapsed

collapse

name of the newly created level

name

a name for the new variable

Value

the original dataframe containing a new column of the collapsed variable with tidyverse code attached

See also

Author

Owen Jin

Examples

collapsed <- collapseLevels(iris, var = "Species",
    levels = c("setosa", "virginica"))
cat(code(collapsed))
#> iris %>% tibble::add_column(Species.coll = forcats::fct_collapse(iris$Species,      setosa_virginica = c("setosa", "virginica")), .after = "Species")
head(collapsed)
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species     Species.coll
#> 1          5.1         3.5          1.4         0.2  setosa setosa_virginica
#> 2          4.9         3.0          1.4         0.2  setosa setosa_virginica
#> 3          4.7         3.2          1.3         0.2  setosa setosa_virginica
#> 4          4.6         3.1          1.5         0.2  setosa setosa_virginica
#> 5          5.0         3.6          1.4         0.2  setosa setosa_virginica
#> 6          5.4         3.9          1.7         0.4  setosa setosa_virginica